Implement more arithmetic operations

+++ tests/tools.py
@@ -4,27 +4,43 @@ import subprocess
 import tempfile
 
 
-def run_compile(source_path, wat_path, wasm_path):
-    res = subprocess.run(f"C:/sw/wasmtime-v33.0.0-x86_64-windows/wasmtime.exe run --dir {source_path.parent.as_posix()} ./build/capylang.wasm -i {source_path.as_posix()} -o {wat_path.as_posix()}", capture_output=True)
-    res = subprocess.run(f"C:/sw/wasm-tools-1.230.0-x86_64-windows/wasm-tools.exe parse {wat_path.as_posix()} -o {wasm_path.as_posix()}", capture_output=True)
+def run_compile(source_path, wat_path, wasm_path) -> tuple[int, str]:
+    res = subprocess.run(
+        f"C:/sw/wasmtime-v33.0.0-x86_64-windows/wasmtime.exe run --dir {source_path.parent.as_posix()} ./build/capylang.wasm -i {source_path.as_posix()} -o {wat_path.as_posix()}",
+        capture_output=True,
+    )
+    if res.returncode != 0:
+        return res.returncode, res.stderr.decode()
+    res = subprocess.run(
+        f"C:/sw/wasm-tools-1.230.0-x86_64-windows/wasm-tools.exe parse {wat_path.as_posix()} -o {wasm_path.as_posix()}",
+        capture_output=True,
+    )
+    return res.returncode, res.stderr.decode()
 
 
 def run_compile_error(source_path, wat_path) -> tuple[int, str]:
-    res = subprocess.run(f"C:/sw/wasmtime-v33.0.0-x86_64-windows/wasmtime.exe run --dir {source_path.parent.as_posix()} ./build/capylang.wasm -i {source_path.as_posix()} -o {wat_path.as_posix()}", capture_output=True)
+    res = subprocess.run(
+        f"C:/sw/wasmtime-v33.0.0-x86_64-windows/wasmtime.exe run --dir {source_path.parent.as_posix()} ./build/capylang.wasm -i {source_path.as_posix()} -o {wat_path.as_posix()}",
+        capture_output=True,
+    )
     return res.returncode, res.stderr.decode()
 
 
 def run_wasmfile(wasm_path) -> tuple[int, str]:
-    res = subprocess.run(f"C:/sw/wasmtime-v33.0.0-x86_64-windows/wasmtime.exe run {wasm_path}", capture_output=True)
-    return res.returncode, res.stdout
+    res = subprocess.run(
+        f"C:/sw/wasmtime-v33.0.0-x86_64-windows/wasmtime.exe run {wasm_path}",
+        capture_output=True,
+    )
+    return res.returncode, res.stdout.decode()
 
 
 def run_test_code(source_code) -> tuple[int, str]:
     """Compiles the given source_code into a WASM module
     and run it in a WASM runtime. The returned values
-    represent the exit code of the run code and the 
+    represent the exit code of the run code and the
     stdout for running it."""
-    source_file = tempfile.NamedTemporaryFile(mode='w', encoding='utf-8', delete=False)
+    source_file = tempfile.NamedTemporaryFile(mode="w", encoding="utf-8",
+                                              delete=False)
     source_file.write(source_code)
     source_file.close()
 
@@ -32,7 +48,10 @@ def run_test_code(source_code) -> tuple[int, str]:
     wat_file_path = source_file_path.with_suffix(".wat")
     wasm_file_path = source_file_path.with_suffix(".wasm")
 
-    run_compile(source_file_path, wat_file_path, wasm_file_path)
+    exit_code, stderr = run_compile(source_file_path, wat_file_path,
+                                    wasm_file_path)
+    if exit_code != 0:
+        raise ValueError("Compilation unexpectedly failed:\n" + stderr)
     exit_code, stdout = run_wasmfile(wasm_file_path)
 
     os.remove(source_file.name)
@@ -49,7 +68,7 @@ def compile_test_code(source_code) -> tuple[int, str]:
     The returned values represent the exit code of the
     compile execution including the run code and the
     stderr."""
-    source_file = tempfile.NamedTemporaryFile(mode='w', encoding='utf-8',
+    source_file = tempfile.NamedTemporaryFile(mode="w", encoding="utf-8",
                                               delete=False)
     source_file.write(source_code)
     source_file.close()
